requirements.txt
pandas==2.2.3
kafka-python==2.1.2
requests==2.32.3
